+++ packages/manager/src/dev-tools/DevTools.tsx
@@ -1,5 +1,4 @@
 import { CloseIcon } from '@linode/ui';
-import { getRoot } from '@linode/utilities';
 import Handyman from '@mui/icons-material/Handyman';
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import { styled } from '@mui/material';
@@ -18,7 +17,6 @@ import { isMSWEnabled } from './utils';
 
 import type { QueryClient } from '@tanstack/react-query';
 import type { ApplicationStore } from 'src/store';
-import { createRoot } from 'react-dom/client';
 
 export type DevToolsView = 'design-tokens' | 'mocks' | 'react-query';
 
@@ -28,7 +26,12 @@ const reactQueryDevtoolsStyle = {
   width: '100%',
 };
 
-const DevTools = () => {
+interface Props {
+  queryClient: QueryClient;
+  store: ApplicationStore;
+}
+
+export const DevTools = (props: Props) => {
   const [isOpen, setIsOpen] = React.useState<boolean>(false);
   const [isDraggable, setIsDraggable] = React.useState<boolean>(false);
   const [view, setView] = React.useState<DevToolsView>('mocks');
@@ -87,118 +90,102 @@ const DevTools = () => {
   }, []);
 
   return (
-    <Draggable draggable={isDraggable}>
-      <div
-        className={`dev-tools ${isMSWEnabled ? 'dev-tools--msw' : ''} ${
-          isOpen ? 'dev-tools--open' : ''
-        } ${isOpen && isDraggable ? 'isDraggable' : ''}
+    <Provider store={props.store}>
+      <Draggable draggable={isDraggable}>
+        <div
+          className={`dev-tools ${isMSWEnabled ? 'dev-tools--msw' : ''} ${
+            isOpen ? 'dev-tools--open' : ''
+          } ${isOpen && isDraggable ? 'isDraggable' : ''}
           `.trim()}
-      >
-        {!isDraggable && (
-          <div className="dev-tools__toggle">
-            <button
-              className="dev-tools-button"
-              onClick={() => setIsOpen(!isOpen)}
-            >
-              <Handyman />
-            </button>
-          </div>
-        )}
-        {isOpen && (
-          <div className="dev-tools__draggable-toggle">
-            <button
-              className="dev-tools-button"
-              onClick={handleDraggableToggle}
-              title="Toggle draggable"
-            >
-              {isDraggable ? <CloseIcon /> : <OpenInNewIcon />}
-            </button>
-          </div>
-        )}
-        <div className="dev-tools__body">
-          <div className="dev-tools__content">
-            <div className="dev-tools__status-bar">
-              <div>
-                <EnvironmentToggleTool />
-              </div>
-              <div className="dev-tools__segmented-button">
-                <button
-                  className={`toggle-button dev-tools-button ${
-                    view === 'mocks' && 'toggle-button--on'
-                  }`}
-                  onClick={handleOpenMocks}
-                >
-                  Mocks
-                </button>
-                <button
-                  className={`toggle-button dev-tools-button ${
-                    view === 'react-query' && 'toggle-button--on'
-                  }`}
-                  onClick={handleOpenReactQuery}
-                >
-                  React Query
-                </button>
-                <button
-                  className={`toggle-button dev-tools-button ${
-                    view === 'design-tokens' && 'toggle-button--on'
-                  }`}
-                  onClick={handleOpenDesignTokens}
-                >
-                  Design Tokens
-                </button>
-              </div>
-              <div>
-                <button
-                  className="dev-tools-button"
-                  onClick={handleGoToPreferences}
-                >
-                  Go to Preferences
-                </button>
-              </div>
+        >
+          {!isDraggable && (
+            <div className="dev-tools__toggle">
+              <button
+                className="dev-tools-button"
+                onClick={() => setIsOpen(!isOpen)}
+              >
+                <Handyman />
+              </button>
             </div>
-            <div className="dev-tools__main" ref={devToolsMainRef}>
-              {view === 'mocks' && (
-                <>
-                  <div className="dev-tools__main__column">
-                    <FeatureFlagTool />
-                  </div>
-                  <div className="dev-tools__main__column">
-                    <ServiceWorkerTool />
-                  </div>
-                </>
-              )}
-              {view === 'react-query' && (
-                <StyledReactQueryDevtoolsContainer
-                  style={reactQueryDevtoolsStyle}
-                >
-                  <QueryClientProvider client={queryClient}>
-                    <ReactQueryDevtools initialIsOpen={true} />
-                  </QueryClientProvider>
-                </StyledReactQueryDevtoolsContainer>
-              )}
-              {view === 'design-tokens' && <DesignTokensTool />}
+          )}
+          {isOpen && (
+            <div className="dev-tools__draggable-toggle">
+              <button
+                className="dev-tools-button"
+                onClick={handleDraggableToggle}
+                title="Toggle draggable"
+              >
+                {isDraggable ? <CloseIcon /> : <OpenInNewIcon />}
+              </button>
+            </div>
+          )}
+          <div className="dev-tools__body">
+            <div className="dev-tools__content">
+              <div className="dev-tools__status-bar">
+                <div>
+                  <EnvironmentToggleTool />
+                </div>
+                <div className="dev-tools__segmented-button">
+                  <button
+                    className={`toggle-button dev-tools-button ${
+                      view === 'mocks' && 'toggle-button--on'
+                    }`}
+                    onClick={handleOpenMocks}
+                  >
+                    Mocks
+                  </button>
+                  <button
+                    className={`toggle-button dev-tools-button ${
+                      view === 'react-query' && 'toggle-button--on'
+                    }`}
+                    onClick={handleOpenReactQuery}
+                  >
+                    React Query
+                  </button>
+                  <button
+                    className={`toggle-button dev-tools-button ${
+                      view === 'design-tokens' && 'toggle-button--on'
+                    }`}
+                    onClick={handleOpenDesignTokens}
+                  >
+                    Design Tokens
+                  </button>
+                </div>
+                <div>
+                  <button
+                    className="dev-tools-button"
+                    onClick={handleGoToPreferences}
+                  >
+                    Go to Preferences
+                  </button>
+                </div>
+              </div>
+              <div className="dev-tools__main" ref={devToolsMainRef}>
+                {view === 'mocks' && (
+                  <>
+                    <div className="dev-tools__main__column">
+                      <FeatureFlagTool />
+                    </div>
+                    <div className="dev-tools__main__column">
+                      <ServiceWorkerTool />
+                    </div>
+                  </>
+                )}
+                {view === 'react-query' && (
+                  <StyledReactQueryDevtoolsContainer
+                    style={reactQueryDevtoolsStyle}
+                  >
+                    <QueryClientProvider client={props.queryClient}>
+                      <ReactQueryDevtools initialIsOpen={true} />
+                    </QueryClientProvider>
+                  </StyledReactQueryDevtoolsContainer>
+                )}
+                {view === 'design-tokens' && <DesignTokensTool />}
+              </div>
             </div>
           </div>
         </div>
-      </div>
-    </Draggable>
-  );
-};
-
-export const install = (store: ApplicationStore, queryClient: QueryClient) => {
-  const devToolsRoot =
-    document.getElementById('dev-tools-root') ||
-    (() => {
-      const newRoot = document.createElement('div');
-      newRoot.id = 'dev-tools-root';
-      document.body.appendChild(newRoot);
-      return newRoot;
-    })();
-
-  const root = createRoot(devToolsRoot);
-  root.render(
-    <Provider store={store}>
-      <DevTools />
+      </Draggable>
     </Provider>
   );
 };
